Fix named export limitation of page files
